🏗️ Issue #1168 パーサー責任分離・適正サイズ化 Phase 2 完了
パーサーファイルの責任分離による保守性向上:

## 主要成果
- 4つの大規模パーサーを責任分離によりモジュール化
- 全ファイルを500行以下に適正サイズ化
- 単一責任原則の徹底適用
- テスト100%通過を維持

## ファイル分割実績
- content_parser.py: 818行 → 3ファイル (231/287/311行)
- list_parser.py: 747行 → 3ファイル (239/358/279行)
- keyword_parser.py: 703行 → 3ファイル (329/297/253行)
- block_parser.py: 661行 → 3ファイル (295/336/313行)

## 責任分離設計
- parser.py: メインクラス・プロトコル準拠・統合制御
- handlers.py: 各種処理ハンドラー・ビジネスロジック
- utils.py: ユーティリティ・パターンマッチング・キャッシュ

## 後方互換性
- レガシーラッパーによる完全互換性維持
- 既存APIの変更なし
- プロトコル準拠の継続
