+++ kumihan_formatter/parsers/block_parser_legacy.py
@@ -1,0 +1,3 @@
+# Legacy wrapper
+from .block.block_parser import UnifiedBlockParser
+BlockParser = UnifiedBlockParser
