+++ kumihan_formatter/parsers/content/__init__.py
@@ -1,0 +1,11 @@
+"""Content Parser Module - コンテンツ解析専用パーサー
+
+責任分離により以下の構造で分割:
+- content_parser.py: メインクラスとパブリックインターフェース
+- content_handlers.py: コンテンツ処理・分析ロジック
+- content_utils.py: ユーティリティ関数群
+"""
+
+from .content_parser import UnifiedContentParser
+
+__all__ = ["UnifiedContentParser"]
