+++ kumihan_formatter/parsers/keyword_parser_legacy.py
@@ -1,0 +1,3 @@
+# Legacy wrapper
+from .keyword.keyword_parser import UnifiedKeywordParser
+KeywordParser = UnifiedKeywordParser
