+++ kumihan_formatter/parsers/list/__init__.py
@@ -1,0 +1,11 @@
+"""List Parser Module - リスト解析専用パーサー
+
+責任分離により以下の構造で分割:
+- list_parser.py: メインクラスとパブリックインターフェース
+- list_handlers.py: リスト処理・分析ロジック
+- list_utils.py: ユーティリティ関数群
+"""
+
+from .list_parser import UnifiedListParser
+
+__all__ = ["UnifiedListParser"]
